Merge borre el aplication.log

+++ sihci/protected/controllers/SponsorsController.php
@@ -174,7 +174,7 @@ class SponsorsController extends Controller {
 
 
 		if (isset($_POST['values1'])) {
-			echo "entre 1";
+			
 			$id_sponsor = Sponsors::model()->findByAttributes(array('id_user' => Yii::app()->user->id))->id;
 			$types = $_POST['types'];
 			$values1 = $_POST['values1'];
@@ -204,7 +204,7 @@ class SponsorsController extends Controller {
 		$id_sponsor = Sponsors::model()->findByAttributes(array("id_user" => Yii::app()->user->id))->id;
 		$model = new SponsorsContacts;
 		$fullname = SponsorsContacts::model()->findAllByAttributes(array("id_sponsor"=>$id_sponsor));
-			// Uncomment the following line if AJAX validation is needed
+		// Uncomment the following line if AJAX validation is needed
 		$this->performAjaxValidation($model);
 		
 		if (isset($_POST['fullnamesUpdate'])) {
@@ -254,6 +254,11 @@ class SponsorsController extends Controller {
 		));
 	}
 
+		public function actionfillFirst() {
+		
+		$this->render('fillFirst');
+	}
+
 	public function actionCreate_billing() {
 
 		$billingExist = SponsorBilling::model()->findByAttributes(array('id_sponsor' => Sponsors::model()->findByAttributes(array('id_user' => Yii::app()->user->id))->id));
@@ -277,14 +282,18 @@ class SponsorsController extends Controller {
 			$model->id_sponsor = $sponsor->id;
 
 			if (isset($_POST['sameAddress'])) {
-				echo "es la misma direccion";
+				
 				$model->id_address_billing = Sponsors::model()->findByAttributes(array("id_user" => Yii::app()->user->id))->id_address;
-				if ($model->save()) {
-					if ($modelAddresses->id != $model->id_address_billing) {
-						if ($modelAddresses->delete());
+				if ($model->save())
+					if ($modelAddresses->id != $model->id_address_billing && $modelAddresses->id > 0) {
+						///echo "if  after model save and same address";
+						if ($modelAddresses->delete())
+							$this->redirect(array('create_billing'));
+					}else
 						$this->redirect(array('create_billing'));
-					} else {
-						echo "no es la misma direccion";
+				} else {
+						//echo "else no same addreres";
+
 						$modelAddresses = new Addresses;
 
 						$modelAddresses->attributes = $_POST['Addresses'];
@@ -292,25 +301,21 @@ class SponsorsController extends Controller {
 						if ($modelAddresses->save()) {
 							$model->id_address_billing = $modelAddresses->id;
 
-							if ($model->save()) {
+							if ($model->save()) 
 								$this->redirect(array('create_billing'));
-							}
+							
 
 						}
 					}
 				}
 
-			}
+			
 
-		}
+		
 
 		$this->render('create_billing', array(
 			'model' => $model, 'modelAddresses' => $modelAddresses, 'sameAd' => $sameAd,
 		));
-	}
-
-	public function actionFillFirst(){
-		$this->render('fillFirst');
 	}
 
 	public function actionCreate_docs() {
@@ -339,7 +344,7 @@ class SponsorsController extends Controller {
 			if (is_object(CUploadedFile::getInstanceByName('Doc1'))) {
 				unset($model);
 				if (!array_key_exists('Documento_que_acredite_la_creacion_de_la_empresa', $modelDocs)) {
-					var_dump($modelDocs);
+					//var_dump($modelDocs);
 					$model = new SponsorsDocs;
 				} else {
 					$model = SponsorsDocs::model()->findByPk($modelDocs['Documento_que_acredite_la_creacion_de_la_empresa'][0]);
@@ -506,6 +511,7 @@ class SponsorsController extends Controller {
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if (!isset($_GET['ajax'])) {
+
 			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
 		}
 
@@ -517,7 +523,8 @@ class SponsorsController extends Controller {
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if (!isset($_GET['ajax'])) {
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			$this->redirect(array('create_contacts'));
+			//$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
 		}
 
 	}
@@ -528,7 +535,8 @@ class SponsorsController extends Controller {
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if (!isset($_GET['ajax'])) {
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			$this->redirect(array('create_contact'));
+			//$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
 		}
 
 	}
